chore: 사용되지 않는 import 문 제거, Controller 메서드명 변경
